Add GauntletEntity

+++ src/main/java/com/cerbon/bosses_of_mass_destruction/config/mob/GauntletConfig.java
@@ -6,7 +6,7 @@ public class GauntletConfig {
     public final double health = 250.0;
     public final double armor = 8.0;
     public final double attack = 16.0;
-    public final double idleHealingPerTick = 0.5f;
+    public final float idleHealingPerTick = 0.5f;
 
     @ConfigEntry.BoundedDiscrete(min = 0, max = 10000)
     public final int experienceDrop = 1000;
